pyproject.toml
[project]
name = "route_lw1_at"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "babel>=2.17.0",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "osmium>=4.2.0",
    "protobuf>=6.33.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
